connections.py: Adds compatibility parameter to utime() call
In order to update the mtime of a timestamp file, the os.utime()
function is called. In py3 variants, simply calling "os.utime(<file>)"
is enough to update the mtime. However, in py2.7, you need to also pass
None as a parameter for it to set the mtime as the current system time.
As this also works in all py3 versions there is no need to check
versions for compatibility.

X-Gentoo-Bug: 569228
X-Gentoo-Bug-URL: https://bugs.gentoo.org/569228